% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yaml_writer.R
\name{count_questions}
\alias{count_questions}
\title{Count number of questions in current lesson}
\usage{
count_questions()
}
\description{
Count number of questions in current lesson
}
\examples{
\dontrun{
count_questions()
}
}

