% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yaml_writer.R
\name{demo_lesson}
\alias{demo_lesson}
\title{Demo the current lesson in swirl}
\usage{
demo_lesson(from = NULL, to = NULL)
}
\arguments{
\item{from}{Question number to begin with. Defaults to beginning of lesson.}

\item{to}{Question number to end with. Defaults to end of lesson.}
}
\description{
Demo the current lesson in swirl
}
\examples{
\dontrun{
# Demo current lesson from beginning through the end
demo_lesson()
# Demo current lesson from question 5 through the end
demo_lesson(5)
# Demo current lesson from question 8 through question 14
demo_lesson(8, 14)
}
}

