% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{find_questions}
\alias{find_questions}
\title{Get question numbers for any questions matching a regular expression}
\usage{
find_questions(regex)
}
\arguments{
\item{regex}{The regular expression to look for in the lesson.
This gets passed along to \code{stringr::str_detect}, so the
same rules apply. See \code{\link[stringr]{str_detect}}.}
}
\value{
A vector of integers representing question numbers.
}
\description{
Get question numbers for any questions matching a regular expression
}
\examples{
\dontrun{
set_lesson()
find_questions("plot")
find_questions("which")
}
}

