% Generated by roxygen2 (4.0.1): do not edit by hand
\name{find_units}
\alias{find_units}
\title{Get unit numbers for any units matching a regular expression}
\usage{
find_units(regex)
}
\arguments{
\item{regex}{The regular expression to look for in the lesson.
This gets passed along to \code{stringr::str_detect}, so the
same rules apply.}
}
\description{
Get unit numbers for any units matching a regular expression
}

