% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_writer.R
\name{get_current_lesson}
\alias{get_current_lesson}
\title{See what lesson you are currently working on}
\usage{
get_current_lesson()
}
\description{
Prints the current lesson and course that you are working on to the console
}
\examples{
\dontrun{
get_current_lesson()
}
}

