% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swirl2html.R
\name{lesson_to_html}
\alias{lesson_to_html}
\title{Turn a swirl lesson into a pretty webpage}
\usage{
lesson_to_html(dest_dir = NULL, open_html = FALSE, keep_rmd = FALSE,
  quiet = FALSE, install_course = TRUE)
}
\arguments{
\item{dest_dir}{destination directory (i.e. where to put the output files).
If not set, default is the lesson directory.}

\item{open_html}{should the html file produced be opened in your browser?
Default is \code{FALSE}.}

\item{keep_rmd}{should the Rmd file be kept after the html is
is produced? Default is \code{FALSE}.}

\item{quiet}{should the rmd rendering output be silenced? Default
is \code{FALSE}.}

\item{install_course}{This is for internal use only. Should the course
be installed? Default is \code{TRUE}.}
}
\description{
At present, this function only accepts a swirl lesson
formatted in YAML. It detects the lesson you are working on
automatically via \code{getOption('swirlify_lesson_file_path')},
converts it to R Markdown (Rmd), then generates a stylized html
document and opens it in your default browser. To prevent clutter,
the Rmd files are not kept by default, but they can be kept
by setting \code{keep_rmd = TRUE}.
}
\details{
The output is formatted to be a readable, standalone tutorial.
This means that information contained in the swirl lesson such as
answer tests and hints are excluded from the Rmd/html output.
}

