% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_writer.R
\name{lp}
\alias{lp}
\title{Get lesson path}
\usage{
lp()
}
\value{
A string, the path to the current \code{lesson.yaml} file.
}
\description{
Find the path to the \code{lesson.yaml} file you're working on.
}
\examples{
\dontrun{
lp()
}
}

