% Generated by roxygen2 (4.0.0): do not edit by hand
\name{make_pathname}
\alias{make_pathname}
\title{Takes a plain English name and turns it into a more proper
file/directory name}
\usage{
make_pathname(name)
}
\description{
Takes a plain English name and turns it into a more proper
file/directory name
}

