% Generated by roxygen2 (4.0.1): do not edit by hand
\name{new_yaml}
\alias{new_yaml}
\title{Create new yaml lesson using yaml_writer}
\usage{
new_yaml(lesson_name, course_name)
}
\arguments{
\item{lesson_name}{Name of lesson}

\item{course_name}{Name of course}
}
\description{
Create new yaml lesson using yaml_writer
}

