% Generated by roxygen2 (4.0.1): do not edit by hand
\name{newYaml}
\alias{newYaml}
\title{Create new yaml lesson using yamlWriter}
\usage{
newYaml(lessonName, courseName)
}
\arguments{
\item{lessonName}{Name of lesson}

\item{courseName}{Name of course}
}
\description{
Create new yaml lesson using yamlWriter
}

