% Generated by roxygen2 (4.0.0): do not edit by hand
\name{setMeta}
\alias{setMeta}
\title{Save the lesson YAML to options.}
\usage{
setMeta(lesson, course, author, organization)
}
\arguments{
\item{lesson}{lesson name}

\item{course}{course name}

\item{author}{your name}

\item{organization}{your organization's name}
}
\description{
This function is for internal use only, but is made public
so that it is accessible from the content authoring app.
}

