% Generated by roxygen2 (4.0.1): do not edit by hand
\name{swirl2html}
\alias{course2html}
\alias{swirl2html}
\title{Turn a swirl lesson into a pretty webpage}
\usage{
swirl2html(dest_dir = NULL, open_html = FALSE, keep_rmd = FALSE)

course2html(course_dir = NULL, dest_dir = NULL, open_html = FALSE,
  keep_rmd = FALSE)
}
\arguments{
\item{dest_dir}{destination directory (i.e. where to put the output files).
If not set, default is the lesson directory.}

\item{open_html}{should the html file produced be opened in your browser?
Default is \code{FALSE}.}

\item{keep_rmd}{should the Rmd file be kept after the html is
is produced? Default is \code{FALSE}.}

\item{course_dir}{path to course directory. If none is specified,
default is the course directory for the lesson you are currently
working on.}
}
\description{
Note that \code{swirl2html()} is an experimental feature. This
it is subject to change based on our own experience and the
feedback we receive from course authors like you.
}
\details{
At present, this function only accepts a swirl lesson
formatted in YAML. It detects the lesson you are working on
automatically via \code{getOption('swirlify_lesson_file_path')},
converts it to R Markdown (Rmd), then generates a stylized html
document and opens it in your default browser. To prevent clutter,
the Rmd files are not kept by default, but they can be kept
by setting \code{keep_rmd = TRUE}.

The output is formatted to be a readable, standalone tutorial.
This means that information contained in the swirl lesson such as
answer tests and hints are excluded from the Rmd/html output.
}

