% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{swirl_courses_dir}
\alias{swirl_courses_dir}
\title{Find the directory where swirl courses are stored}
\usage{
swirl_courses_dir()
}
\value{
A string with the path to where swirl is searching for courses.
}
\description{
Find the directory where swirl courses are stored
}
