% Generated by roxygen2 (4.0.0): do not edit by hand
\name{swirlify}
\alias{swirlify}
\title{Author lesson using content authoring tool}
\usage{
swirlify(course, lesson)
}
\arguments{
  \item{course}{course name}

  \item{lesson}{lesson name}
}
\description{
Author lesson using content authoring tool
}

