% Generated by roxygen2 (4.0.1): do not edit by hand
\name{test_lesson_by_name}
\alias{test_lesson_by_name}
\title{Test all cmd questions of any lesson of current course.}
\usage{
test_lesson_by_name(lesson_dir_name)
}
\arguments{
\item{lesson_dir_name}{}
}
\description{
Test all cmd questions of any lesson of current course.
}

