% Generated by roxygen2 (4.0.1): do not edit by hand
\name{testit}
\alias{testit}
\title{Test the current lesson in swirl}
\usage{
testit(from = NULL, to = NULL)
}
\arguments{
\item{from}{Unit number to begin with. Defaults to beginning of lesson.}

\item{to}{Unit number to end with. Defaults to end of lesson.}
}
\description{
Test the current lesson in swirl
}
\examples{
\dontrun{
# Test current lesson from beginning through end
testit()
# Test current lesson from unit 5 through end
testit(5)
# Test current lesson from unit 8 through unit 14
testit(8, 14)
}
}

