% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wq.R
\name{wq_script}
\alias{wq_script}
\title{Template for R script question}
\usage{
wq_script(output = "explain what the user must do here",
  answer_tests = "custom_test_name()", hint = "hint",
  script = "script-name.R")
}
\arguments{
\item{output}{Text that is displayed to the user.}

\item{answer_tests}{An internal function from \code{swirl} for testing the 
user's choice. See \code{\link[swirl]{AnswerTests}}.}

\item{hint}{A string that is printed to the console if the user answers this
question incorrectly.}

\item{script}{The name of the script template to be opened. This template
should be in a directory called \code{scripts} located inside the lesson
directory.}
}
\description{
Template for R script question
}
\examples{
\dontrun{
# While writing a new lesson by hand just use:
wq_script()

# If converting from another format to a swirl course you may want to sue the
# API:
wq_script("Write a function that adds three numbers.",
 "add_three_test()", "Something like: add3 <- function(x, y, z){x+y+z}", 
 "add-three.R")
}
}

