% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_writer.R
\name{wq_video}
\alias{wq_video}
\title{Template for video question}
\usage{
wq_video(output = "Would you like to watch a short video about ___?",
  video_link = "http://address.of.video")
}
\arguments{
\item{output}{Text that is displayed to the user.}

\item{video_link}{A link to a url. Please make sure to use \code{http://} or
\code{https://}.}
}
\description{
The \code{url} provided for \code{video_link} can be a link to any website.
}
\examples{
\dontrun{
# While writing a new lesson by hand just use:
wq_video()

# If converting from another format to a swirl course you may want to sue the
# API:
wq_video("Now Roger will show you the basics on YouTube.",
 "https://youtu.be/dQw4w9WgXcQ")
}
}

