% Generated by roxygen2 (4.0.0): do not edit by hand
\name{write_unit}
\alias{write_unit}
\title{TODO:
- Checks to prevent overwriting existing content
- Migrate course authoring tools from swirl package
- Integrate with existing functions (e.g. author_lesson())
- Write new menu() methods for more efficient testing of new content
- Throw helpful errors (if missing function args, etc)}
\usage{
write_unit(lessonFile)
}
\description{
TODO:
- Checks to prevent overwriting existing content
- Migrate course authoring tools from swirl package
- Integrate with existing functions (e.g. author_lesson())
- Write new menu() methods for more efficient testing of new content
- Throw helpful errors (if missing function args, etc)
}

