% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{count_questions}
\alias{count_questions}
\title{Count number of questions in current lesson}
\usage{
count_questions()
}
\value{
Number of questions as an integer, invisibly
}
\description{
Returns and prints the number of questions in the current lesson.
}
\examples{
\dontrun{
count_questions()
}
}

