% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{make_pathname}
\alias{make_pathname}
\title{Replace spaces in strings with underscores}
\usage{
make_pathname(name)
}
\arguments{
\item{name}{A vector of strings.}
}
\value{
A string vector where spaces are replaced with underscores.
}
\description{
Useful for creating paths to a particular swirl course, as you might want
to do in files like \code{initLesson.R}.
}
\examples{
make_pathname("Developing Data Products")
# "Developing_Data_Products"

make_pathname(c("R Programming", "Exploratory Data Analysis"))
# "R_Programming" "Exploratory_Data_Analysis"
}
