% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yaml_writer.R
\name{new_lesson}
\alias{new_lesson}
\title{Create new lesson in the YAML format.}
\usage{
new_lesson(lesson_name, course_name)
}
\arguments{
\item{lesson_name}{Name of lesson}

\item{course_name}{Name of course}
}
\description{
This is the recommended method for creating new content.
}

