% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yaml_writer.R
\name{set_lesson}
\alias{set_lesson}
\title{Select an existing lesson you want to work on}
\usage{
set_lesson(path_to_yaml = NULL, open_lesson = TRUE, silent = FALSE)
}
\arguments{
\item{path_to_yaml}{Optional, full path to YAML lesson file. If not
specified, then you will be prompted to select file interactively.}

\item{open_lesson}{Should the lesson be opened automatically?
Default is \code{TRUE}.}

\item{silent}{Should the lesson be set silently? Default is
\code{FALSE}.}
}
\description{
Select an existing lesson you want to work on
}
\examples{
\dontrun{
# Set the lesson interactively
set_lesson()

# You can also specify the path to the \\code{yaml} file you wish to work on.
set_lesson(file.path("~", "R_Programming", "Functions", "lesson.yaml"))
}
}

