% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set_lesson}
\alias{set_lesson}
\title{Select an existing lesson you want to work on}
\usage{
set_lesson(path2yaml = NULL)
}
\arguments{
\item{path2yaml}{Optional, full path to YAML lesson file. If not
specified, then you will be prompted to select file interactively.}
}
\description{
Select an existing lesson you want to work on
}

