% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{swirl2html}
\alias{swirl2html}
\title{Turn a swirl lesson into a pretty webpage}
\usage{
swirl2html(destDir)
}
\arguments{
\item{destDir}{destination directory (i.e. where to put the R
markdown (Rmd) and html output files)}
}
\description{
Note that \code{swirl2html()} is an experimental feature. This
it is subject to change based on our own experience and the
feedback we receive from course authors like you.
}
\details{
At present, this function only accepts a swirl lesson
formatted in YAML. It detects the lesson you are working on
automatically via \code{getOption('swirlify_lesson_file_path')},
converts it to R Markdown (Rmd), then generates a stylized html
document and opens it in your default browser.

The output is formatted to be a readable, standalone tutorial.
This means that information contained in the swirl lesson such as
answer tests and hints are excluded from the Rmd/html output.
}

