% Generated by roxygen2 (4.0.0): do not edit by hand
\name{swirlify}
\alias{swirlify}
\title{Author lesson using content authoring app.}
\usage{
swirlify(lesson, course, author = NULL, organization = NULL)
}
\arguments{
\item{lesson}{lesson name}

\item{course}{course name}

\item{author}{your name}

\item{organization}{your organization's name}
}
\description{
Author lesson using content authoring app.
}
\examples{
\dontrun{

swirlify("Linear Regression", Statistics 101")
}
}

