% Generated by roxygen2 (4.0.0): do not edit by hand
\name{swirlify}
\alias{swirlify}
\title{Author lesson using content authoring tool}
\usage{
swirlify(lesson, course, lesson_type = "yaml", where = NULL)
}
\arguments{
  \item{lesson}{Lesson name}

  \item{course}{Course name}

  \item{lesson_type}{Lesson type. At present, must be
  either "yaml" (default) or "rmd".}

  \item{where}{Directory within which course directory will
  be created. By default, your current working directory.}
}
\description{
Author lesson using content authoring tool
}
\examples{
\dontrun{

swirlify("Linear Regression", "Statistics 101", "yaml")
}
}

