% Generated by roxygen2 (4.0.0): do not edit by hand
\name{swirlify}
\alias{swirlify}
\title{Author lesson using content authoring tool}
\usage{
swirlify(lesson, course)
}
\arguments{
  \item{lesson}{lesson name}

  \item{course}{course name}
}
\description{
Author lesson using content authoring tool
}
\examples{
\dontrun{

swirlify("Linear Regression", Statistics 101")
}
}

