% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_lesson.R
\name{test_lesson}
\alias{test_lesson}
\title{Test all cmd and mult questions of a lesson.}
\usage{
test_lesson(lesson_dir_name = NULL)
}
\arguments{
\item{lesson_dir_name}{The directory name of the lesson. Defaults to current lesson.}
}
\description{
Test all cmd and mult questions of a lesson.
}

