% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_lesson.R
\name{test_lesson}
\alias{test_lesson}
\title{Run tests on a lesson}
\usage{
test_lesson()
}
\description{
Run basic tests on all questions in the current lesson.
}
\examples{
\dontrun{
# Set a lesson interactively
set_lesson()

# Run tests on that lesson
test_lesson()
}
}

