% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yaml_writer.R
\name{test_lesson}
\alias{test_lesson}
\title{Test the current lesson in swirl}
\usage{
test_lesson(from = NULL, to = NULL)
}
\arguments{
\item{from}{Unit number to begin with. Defaults to beginning of lesson.}

\item{to}{Unit number to end with. Defaults to end of lesson.}
}
\description{
Test the current lesson in swirl
}
\examples{
\dontrun{
# Test current lesson from beginning through end
test_lesson()
# Test current lesson from unit 5 through end
test_lesson(5)
# Test current lesson from unit 8 through unit 14
test_lesson(8, 14)
}
}

