<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_lesson.R
\name{test_lesson}
\alias{test_lesson}
\title{Run tests on a lesson}
\usage{
test_lesson()
}
\description{
Run basic tests on all questions in the current lesson.
}
\examples{
\dontrun{
# Set a lesson interactively
set_lesson()

# Run tests on that lesson
test_lesson()
}
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test_lesson.R
\name{test_lesson}
\alias{test_lesson}
\title{Test a YAML Lesson File

This function tests whether the \code{CorrectAnswer} matches
the \code{AnswerTests}.

It only checks \code{cmd_question} and \code{mult_question} now.}
\usage{
test_lesson(lesson_dir = NULL)
}
\arguments{
\item{lesson_dir}{path. The location of lesson.}
}
\description{
Test a YAML Lesson File

This function tests whether the \code{CorrectAnswer} matches
the \code{AnswerTests}.

It only checks \code{cmd_question} and \code{mult_question} now.
>>>>>>> 17f3335ef8ea87b0812ccb3e557f9fc6800395db
}

