% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_lesson.R
\name{testit}
\alias{testit}
\title{(Deprecated)}
\usage{
testit(from = NULL, to = NULL)
}
\arguments{
\item{from}{Question number to begin with. Defaults to beginning of lesson.}

\item{to}{Question number to end with. Defaults to end of lesson.}
}
\description{
This function is deprecated. Please use \code{demo_lesson}
instead.
}

